refactor(form-builder): cleanup legacy part support in the PTE
This will remove the part 'part:@sanity/form-builder/input/block-editor/block-extras' but still support
the more poentially used PTE parts. This is undocumented support, but still continue to support the most
likeley use case for overriding the markers parts

This will also rename some files, and make sure we import the actual parts in the PTE code
