+++ packages/@sanity/form-builder/src/inputs/PortableText/_legacyDefaultParts/Markers.tsx
@@ -1,12 +1,9 @@
-// @todo: remove the following line when part imports has been removed from this file
-///<reference types="@sanity/types/parts" />
-
 import React, {useMemo} from 'react'
-import CustomMarkers from 'part:@sanity/form-builder/input/block-editor/block-markers-custom-default'
 import {Marker, isValidationMarker} from '@sanity/types'
 import {Box, Flex, Stack, Text, Theme} from '@sanity/ui'
 import {InfoOutlineIcon, WarningOutlineIcon} from '@sanity/icons'
 import styled, {css} from 'styled-components'
+import {CustomMarkers} from '../../../legacyParts'
 import {RenderCustomMarkers} from '../types'
 
 type Props = {
