--- packages/@sanity/form-builder/src/inputs/PortableText/legacyParts/BlockExtras.module.css
@@ -1,81 +1,0 @@
-@import 'part:@sanity/base/theme/variables-style';
-
-.root {
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-  position: absolute;
-  z-index: var(--zindex-portal);
-  overflow: visible;
-}
-
-.content {
-  height: 100%;
-  margin-left: calc(100% - var(--block-extras-width));
-  border-left: 2px solid transparent;
-  padding: var(--small-padding) 0 var(--small-padding) calc(var(--small-padding) - 2px);
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-left: 100%;
-
-  @nest &:hover::before {
-    content: '';
-    position: absolute;
-    box-sizing: border-box;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: color(var(--gray) a(10%));
-  }
-}
-
-.withError {
-  composes: root;
-
-  @nest & .content {
-    border-color: var(--state-danger-color);
-  }
-
-  @nest & .content:hover::before {
-    background-color: color(var(--state-danger-color) a(10%));
-  }
-}
-
-.withWarning {
-  composes: root;
-  border-color: var(--state-warning-color);
-
-  @nest &:hover::before {
-    background-color: color(var(--state-warning-color) a(10%));
-  }
-}
-
-.blockActions {
-  display: flex;
-  pointer-events: all !important;
-
-  @nest & .content {
-    color: var(--text-color);
-    border-color: var(--gray);
-  }
-}
-
-.markers {
-  display: flex;
-  pointer-events: all !important;
-  width: calc(var(--block-extras-width) - var(--small-padding) * 2);
-
-  @nest & .content {
-    border-color: var(--state-info-color);
-    color: var(--text-color);
-  }
-}
-
-.hasFocus {
-}
-
-.changeIndicator {
-}
