--- packages/@sanity/form-builder/src/inputs/PortableText/legacyParts/Markers.module.css
@@ -1,6 +1,0 @@
-@import 'part:@sanity/base/theme/variables-style';
-
-.markerGroup {
-  display: inline-block;
-  padding-right: var(--small-padding);
-}
