+++ packages/@sanity/form-builder/src/inputs/PortableText/object/BlockObject.tsx
@@ -12,7 +12,7 @@ import React, {useCallback, useMemo, useRef} from 'react'
 import styled, {css} from 'styled-components'
 import {ChangeIndicatorWithProvidedFullPath} from '@sanity/base/change-indicators'
 import {RenderBlockActions, RenderCustomMarkers} from '../types'
-import Markers from '../legacyParts/Markers'
+import {Markers} from '../../../legacyParts'
 import PatchEvent from '../../../PatchEvent'
 import {BlockActions} from '../BlockActions'
 import {BlockObjectPreview} from './BlockObjectPreview'
