+++ packages/@sanity/form-builder/src/inputs/PortableText/text/TextBlock.tsx
@@ -5,8 +5,8 @@ import {hues} from '@sanity/color'
 import {isKeySegment, Marker} from '@sanity/types'
 import {PortableTextBlock, RenderAttributes} from '@sanity/portable-text-editor'
 import {ChangeIndicatorWithProvidedFullPath} from '@sanity/base/change-indicators'
+import {Markers} from '../../../legacyParts'
 import {RenderBlockActions, RenderCustomMarkers} from '../types'
-import Markers from '../legacyParts/Markers'
 import PatchEvent from '../../../PatchEvent'
 import {BlockActions} from '../BlockActions'
 import {
