--- packages/@sanity/form-builder/src/sanity/legacyPartImplementations/block-extras-default.ts
@@ -1,1 +1,0 @@
-export {default} from '../../inputs/PortableText/legacyParts/BlockExtras'
