+++ packages/@sanity/form-builder/src/sanity/legacyPartImplementations/block-markers-custom-default.ts
@@ -1,1 +1,1 @@
-export {default} from '../../inputs/PortableText/legacyParts/CustomMarkers'
+export {default} from '../../inputs/PortableText/_legacyDefaultParts/CustomMarkers'
